requirements.txt
kmeans_pytorch==0.3
numpy==1.23.3
pandas==1.4.4
Pillow==9.5.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.1
wandb==0.13.9
